Move `WolfContext{,Builder}` into `context.rs`
