+++ src/async_client.rs
@@ -26,7 +26,11 @@
 //! [4]: https://www.wolfssl.com/documentation/manuals/wolfssl/group__IO.html#function-wolfssl_write
 //! [5]: https://docs.rs/tokio/latest/tokio/net/struct.TcpStream.html#impl-AsRawFd-for-TcpStream
 
-use crate::{errors::WolfError, session::WolfSession, WolfContext, WolfContextBuilder};
+use crate::{
+    context::{WolfContext, WolfContextBuilder},
+    errors::WolfError,
+    session::WolfSession,
+};
 
 use bytes::{Buf, BytesMut};
 use pin_project::pin_project;
